+++ Models/DeviceHkrNextchange.cs
@@ -5,12 +5,12 @@ namespace Fritz.HomeAutomation.Models
 {
     [Serializable]
     [XmlType(AnonymousType = true)]
-    public  class DeviceHkrNextChange
+    public class DeviceHkrNextChange
     {
         [XmlElement("endperiod")]
         public int EndPeriod { get; set; }
 
         [XmlElement("tchange")]
-        public byte TChange { get; set; }
+        public int TChange { get; set; }
     }
 }
